changed where we test the classes
